Fix formatting of new renderthread message
On 32-bit builds a warning would be generated because of a mistmatch
between size_t and the %lu formatting parameter. This caused Windows
builds to fail. In the unlikely event that we ever go into the billions
of render threads this change will now cast the value to uin64_t and use
the PRIu64 macro to ensure that it works on all platforms.

+++ distrib/android-emugl/host/libs/libOpenglRender/RendererImpl.cpp
@@ -115,8 +115,8 @@ RenderChannelPtr RendererImpl::createRenderChannel() {
                                       }),
                        mThreads.end());
 
-        DBG("Started new RenderThread (total %lu) @%p\n",
-            mThreads.size() + 1, rt.get());
+        DBG("Started new RenderThread (total %" PRIu64 ") @%p\n",
+            static_cast<uint64_t>(mThreads.size() + 1), rt.get());
 
         mThreads.emplace_back(std::move(rt), channel);
     }
